business_module/requirements.txt
django-cors-headers
mysqlclient
Django~=5.0.4
djangorestframework~=3.15.1
djangorestframework-simplejwt
cloudinary
django-cloudinary-storage